refactor discv5 constructs and test it with an eth2 client
