--- devp2p/src/main/kotlin/org/apache/tuweni/devp2p/v5/TopicQueryMessageHandler.kt
@@ -1,43 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.apache.tuweni.devp2p.v5
-
-import org.apache.tuweni.bytes.Bytes
-import org.apache.tuweni.devp2p.v5.topic.Topic
-import java.net.InetSocketAddress
-
-internal class TopicQueryMessageHandler : MessageHandler<TopicQueryMessage> {
-
-  companion object {
-    private const val MAX_NODES_IN_RESPONSE: Int = 16
-  }
-
-  override suspend fun handle(
-    message: TopicQueryMessage,
-    address: InetSocketAddress,
-    srcNodeId: Bytes,
-    connector: UdpConnector
-  ) {
-    val topicTable = connector.getTopicTable()
-    val nodes = topicTable.getNodes(Topic(message.topic.toHexString()))
-
-    nodes.chunked(MAX_NODES_IN_RESPONSE).forEach {
-      val response = NodesMessage(message.requestId, nodes.size, it)
-      connector.send(address, response, srcNodeId)
-    }
-  }
-}
