+++ devp2p/src/test/kotlin/org/apache/tuweni/devp2p/v5/packet/FindNodeMessageTest.kt
@@ -42,6 +42,6 @@ class FindNodeMessageTest {
   fun getMessageTypeHasValidIndex() {
     val message = FindNodeMessage()
 
-    assert(3 == message.getMessageType().toInt())
+    assert(3 == message.messageIdentifier().toInt())
   }
 }
