+++ devp2p/src/test/kotlin/org/apache/tuweni/devp2p/v5/packet/RandomMessageTest.kt
@@ -18,7 +18,7 @@ package org.apache.tuweni.devp2p.v5.packet
 
 import org.apache.tuweni.bytes.Bytes
 import org.apache.tuweni.devp2p.v5.RandomMessage
-import org.apache.tuweni.devp2p.v5.UdpMessage
+import org.apache.tuweni.devp2p.v5.Message
 import org.junit.jupiter.api.Test
 
 class RandomMessageTest {
@@ -29,12 +29,12 @@ class RandomMessageTest {
       "0xb53ccf732982b8e950836d1e02898c8b38cfdbfdf86bc65c8826506b454e14618ea73612a0f5582c130ff666"
 
     val data = Bytes.fromHexString(expectedEncodingResult)
-    val message = RandomMessage(UdpMessage.authTag(), data)
+    val message = RandomMessage(Message.authTag(), data)
 
     val encodingResult = message.encode()
     assert(encodingResult.toHexString() == expectedEncodingResult)
 
-    val decodingResult = RandomMessage.create(UdpMessage.authTag(), encodingResult)
+    val decodingResult = RandomMessage.create(Message.authTag(), encodingResult)
 
     assert(decodingResult.data == data)
   }
@@ -43,7 +43,7 @@ class RandomMessageTest {
   fun randomDataGivesRandom44Bytes() {
     val firstResult = RandomMessage.randomData()
 
-    assert(UdpMessage.RANDOM_DATA_LENGTH == firstResult.size())
+    assert(Message.RANDOM_DATA_LENGTH == firstResult.size())
 
     val secondResult = RandomMessage.randomData()
