+++ devp2p/src/test/kotlin/org/apache/tuweni/devp2p/v5/packet/RegConfirmationMessageTest.kt
@@ -39,6 +39,6 @@ class RegConfirmationMessageTest {
   fun getMessageTypeHasValidIndex() {
     val message = RegConfirmationMessage(topic = Bytes.random(32))
 
-    assert(7 == message.getMessageType().toInt())
+    assert(7 == message.messageIdentifier().toInt())
   }
 }
