+++ devp2p/src/test/kotlin/org/apache/tuweni/devp2p/v5/packet/RegTopicMessageTest.kt
@@ -45,6 +45,6 @@ class RegTopicMessageTest {
       topic = Bytes.random(16)
     )
 
-    assert(5 == message.getMessageType().toInt())
+    assert(5 == message.messageIdentifier().toInt())
   }
 }
