+++ devp2p/src/test/kotlin/org/apache/tuweni/devp2p/v5/packet/TopicQueryMessageTest.kt
@@ -39,6 +39,6 @@ class TopicQueryMessageTest {
   fun getMessageTypeHasValidIndex() {
     val message = TopicQueryMessage(topic = Bytes.random(32))
 
-    assert(8 == message.getMessageType().toInt())
+    assert(8 == message.messageIdentifier().toInt())
   }
 }
